moved tardis app functions out of urls.py
